+++ tests/test_toolbars.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.models import Permission
+from django.utils.text import slugify
 
 from cms.cms_toolbars import LANGUAGE_MENU_IDENTIFIER, PlaceholderToolbar
 from cms.test_utils.testcases import CMSTestCase
@@ -12,6 +13,7 @@ from djangocms_versioning.test_utils.factories import (
     BlogPostVersionFactory,
     FancyPollFactory,
     PageContentWithVersionFactory,
+    PageUrlFactory,
     PageVersionFactory,
     PollVersionFactory,
     UserFactory,
@@ -302,6 +304,132 @@ class VersioningToolbarTestCase(CMSTestCase):
         )
 
         self.assertEqual(expected_label, version_menu.name)
+
+    def test_view_published_in_toolbar_in_edit_mode_for_published_page(self):
+        """
+        The 'View Published' control is only relevant for pages that
+        are published
+        """
+        published_version = PageVersionFactory(content__language="en", state=PUBLISHED)
+        toolbar = get_toolbar(published_version.content, edit_mode=True)
+
+        toolbar.post_template_populate()
+
+        self.assertTrue(toolbar_button_exists("View Published", toolbar.toolbar))
+
+    def test_view_published_in_toolbar_in_preview_mode_for_published_page(self):
+        """
+        The 'View Published' control is only relevant for pages that
+        are published
+        """
+        published_version = PageVersionFactory(content__language="en", state=PUBLISHED)
+        toolbar = get_toolbar(published_version.content, preview_mode=True)
+
+        toolbar.post_template_populate()
+
+        self.assertTrue(toolbar_button_exists("View Published", toolbar.toolbar))
+
+    def test_view_published_not_in_toolbar_in_edit_mode_for_draft_page(self):
+        """
+        The 'View Published' control is only relevant for pages that
+        are published
+        """
+        draft_version = PageVersionFactory(content__language="en")
+        toolbar = get_toolbar(draft_version.content, edit_mode=True)
+
+        toolbar.post_template_populate()
+
+        self.assertFalse(toolbar_button_exists("View Published", toolbar.toolbar))
+
+    def test_view_published_not_in_toolbar_in_preview_mode_for_draft_page(self):
+        """
+        The 'View Published' control is only relevant for pages that
+        are published
+        """
+        draft_version = PageVersionFactory(content__language="en")
+        toolbar = get_toolbar(draft_version.content, preview_mode=True)
+
+        toolbar.post_template_populate()
+
+        self.assertFalse(toolbar_button_exists("View Published", toolbar.toolbar))
+
+    def test_view_published_not_in_toolbar_in_edit_mode_for_poll(self):
+        """
+        The 'View Published' toolbar control is only relevant for pages that have
+        the concept of a live url / web viewable url with the toolbar
+        """
+        version = PollVersionFactory(state=PUBLISHED)
+        toolbar = get_toolbar(version.content, edit_mode=True)
+
+        toolbar.post_template_populate()
+
+        self.assertFalse(toolbar_button_exists("View Published", toolbar.toolbar))
+
+    def test_view_published_not_in_toolbar_in_preview_mode_for_poll(self):
+        """
+        The 'View Published' toolbar control is only relevant for pages that have
+        the concept of a live url / web viewable url with the toolbar
+        """
+        version = PollVersionFactory(state=PUBLISHED)
+        toolbar = get_toolbar(version.content, preview_mode=True)
+
+        toolbar.post_template_populate()
+
+        self.assertFalse(toolbar_button_exists("View Published", toolbar.toolbar))
+
+    def test_view_published_in_toolbar_in_edit_mode_button_url(self):
+        """
+        The 'View Published' toolbar control url should be a valid
+        live url when in the edit mode
+        """
+        published_version = PageVersionFactory(content__language="en")
+        language = published_version.content.language
+        PageUrlFactory(
+            page=published_version.content.page,
+            language=language,
+            path=slugify('test_page'),
+            slug=slugify('test_page'),
+        )
+        published_version.publish(user=self.get_superuser())
+        draft_version = published_version.copy(self.get_superuser())
+        edit_endpoint = get_object_edit_url(draft_version.content)
+        expected_url = published_version.content.page.get_absolute_url(language=language)
+
+        with self.login_user_context(self.get_superuser()):
+            response = self.client.get(edit_endpoint)
+
+        found_button_list = find_toolbar_buttons("View Published", response.wsgi_request.toolbar)
+
+        # check only one View Published button exists
+        self.assertEqual(len(found_button_list), 1)
+        self.assertEqual(found_button_list[0].url, expected_url)
+
+    def test_view_published_in_toolbar_in_preview_mode_button_url(self):
+        """
+        The 'View Published' toolbar control url should be a valid
+        live url when in the preview mode
+        """
+        published_version = PageVersionFactory(content__language="en")
+        language = published_version.content.language
+        PageUrlFactory(
+            page=published_version.content.page,
+            language=language,
+            path=slugify('test_page'),
+            slug=slugify('test_page'),
+        )
+        published_version.publish(user=self.get_superuser())
+        draft_version = published_version.copy(self.get_superuser())
+        preview_endpoint = get_object_preview_url(draft_version.content)
+        expected_url = published_version.content.page.get_absolute_url(language=language)
+
+        with self.login_user_context(self.get_superuser()):
+            response = self.client.get(preview_endpoint)
+
+        found_button_list = find_toolbar_buttons("View Published", response.wsgi_request.toolbar)
+
+        # check only one View Published button exists
+        self.assertEqual(len(found_button_list), 1)
+        self.assertEqual(found_button_list[0].url, expected_url)
 
 
 class VersioningPageToolbarTestCase(CMSTestCase):